ee-map-draw-demo-backend/requirements.txt
google-cloud-bigquery==2.34.4
Flask==2.1.0
earthengine-api==0.1.358
flask-cors==3.0.10
google-cloud-firestore==2.9.0
geojson==3.0.0
shapely==2.0.0
gunicorn==20.1.0
